feat: Create DynamicWhereFilter (#26)
* Create DynamicWhereFilter.php

* Create CustomWhereFilter.php

* Update DynamicWhereFilter.php

* Delete app/Domain/Common/Filters/CustomWhereFilter.php

* Create ColumnTypeCache.php

* Create ComboSearchFilter.php

* style: apply PHP CS Fixer

* Update and rename DynamicWhereFilter.php to AdvancedFilter.php

* Update AdvancedFilter.php

* Create AdvancedFilterTest

* Tidy & refactor

- add user status
- fix filters
- tenant policy

---------

+++ app/Domain/Tenant/Controllers/TenantController.php
@@ -6,12 +6,15 @@ use App\Domain\Tenant\DTOs\TenantDTO;
 use App\Domain\Tenant\Models\Tenant;
 use App\Domain\Tenant\Requests\TenantRequest;
 use App\Http\Controllers\Controller;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 class TenantController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request): JsonResponse
     {
         $tenants = $request->user()->tenants()->orderBy('created_at')->get();
@@ -32,8 +35,10 @@ class TenantController extends Controller
         );
     }
 
-    public function show(Tenant $tenant): JsonResponse
+    public function show(Request $request, Tenant $tenant): JsonResponse
     {
+        $this->authorize('view', $tenant);
+
         return response()->json(
             TenantDTO::from($tenant)
         );
@@ -41,7 +46,7 @@ class TenantController extends Controller
 
     public function update(TenantRequest $request, Tenant $tenant): JsonResponse
     {
-        $tenant = $request->user()->tenants()->where('id', $tenant->id)->firstOrFail();
+        $this->authorize('update', $tenant);
         $tenant->update($request->validated());
 
         return response()->json(
@@ -51,7 +56,7 @@ class TenantController extends Controller
 
     public function destroy(Request $request, Tenant $tenant): JsonResponse
     {
-        $tenant = $request->user()->tenants()->where('id', $tenant->id)->firstOrFail();
+        $this->authorize('delete', $tenant);
         $tenant->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
